feat: moved all the const of the components to constants folder
