+++ src/components/SalariesByExperience.jsx
@@ -1,31 +1,13 @@
-import { IconBuildingSkyscraper, IconSquareHalf, IconBeach } from '@tabler/icons-react'
 import { Card, Metric, Text, BarList, Flex, Grid } from '@tremor/react'
 import { formatNumberToEur } from '@/utils/formatters'
-
-const data = [
-  {
-    category: 'Oficina',
-    field: 'office',
-    icon: <IconBuildingSkyscraper stroke={1} />
-  },
-  {
-    category: 'Híbrido',
-    field: 'hybrid',
-    icon: <IconSquareHalf stroke={1} />
-  },
-  {
-    category: '100% Remoto',
-    field: 'remote',
-    icon: <IconBeach stroke={1} />
-  }
-]
+import { DATA_SALARIES_BY_EXPERIENCE } from '@/constants'
 
 export function SalariesByExperience ({ averageSalaries }) {
   return (
     <>
       <h2 className='text-lg font-medium text-gray-700'>Sueldos por modalidad de trabajo y experiencia</h2>
       <Grid id='salaries-general' numColsSm={2} numColsLg={3} className='gap-6 mt-6'>
-        {data.map((item) => (
+        {DATA_SALARIES_BY_EXPERIENCE.map((item) => (
           <Card key={item.category}>
 
             <header className='flex justify-between'>
